backend/requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
defusedxml==0.7.1
Flask==3.0.3
gunicorn==22.0.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.2.2
MarkupSafe==2.1.5
odfpy==1.4.1
packaging==24.0
pluggy==1.5.0
pytest==8.2.2
python-docx==1.1.2
typing_extensions==4.12.1
Werkzeug==3.0.3
